Rename previous default file to _minirun6 and rename the _refactor version as the default.
